refactor: move all application logic into lib, main is now a thin entry point
